Fixed int => decimal(10,2) in Payment and Room + SP

+++ CoCowork.DB/StoredProcedures/Payment/Payment_Insert.sql
@@ -1,5 +1,5 @@
 ﻿CREATE PROC dbo.Payment_Insert
-	@Amount int,
+	@Amount decimal(10,2),
 	@PaymentDate DateTime,
 	@OrderId int
 AS
