+++ CoCowork.DB/Tables/Payment.sql
@@ -1,7 +1,7 @@
 ﻿CREATE TABLE [dbo].[Payment]
 (
 	[Id] INT NOT NULL PRIMARY KEY IDENTITY,
-	[Amount] int NOT NULL,
+	[Amount] DECIMAL(10, 2) NOT NULL,
 	[PaymentDate] DateTime NOT NULL,
 	[OrderId] int NOT NULL, 
 )
